feat (ecal): Implement new cait.versatile energy calibration interface with modular insertion of different testpulse responses and transfer functions.
Implemented abstract classes 'vai.TestpulseResponse' and 'vai.TransferFunction' and polynomial as well as spline interpolation implementations. More might follow.
Implemented a vai.EnergyCalibration object which combines the TestpulseResponse and TransferFunction objects to perform energy calibration both in the forward, as well as the inverse direction (i.e. TPE->PH and PH->TPE). Also implemented an interactive widget to test the workings of the calibration in vai.EnergyCalibration.preview()