perf(MainParameters): Slight improvement in speed
MainParameters now avoids the loop over channels by using `numpy.argmax`
on an array of `bool`s (a mask array), which leads to ~10% improvement
in run time compared to using the loop.

This method might also make it unnecessary to reshape the event, but
that also needs to be tested and the output shape fixed.