feat+fix: Enforce peak_loc properly, and allow lists of ints
The `peak_loc` argument was not previously doing anything if it was a
list, only if it were an int or float.  Now it properly sets bounds on
the search region for the peak position.

Also now allows lists of integers to be passed to `peak_loc` argument,
which sets the bounds of the search region in samples.  Checks that
these are inside the range of an event are impossible though, as there
is no way to know the shape of the incoming event(s) in the constructor.