fix: Calculation of BL RMS/slope when model is 0
The baseline RMS returned from FitBaseline when the model is not a
0-order polynomial is actually the root-sum-square, but if the model is
0th order then np.std is used, which is the root-mean-square.
Previously, the value returned from FitBaseline was divided by sqrt(N)
regardless of the model, which should not be used if the model is 0.

Additionally, if the model is 0, then the slope of the trace is
approximated from the endpoints of the `where` argument to FitBaseline.