bug: Add axis if single channel is processed
For backwards compatibility (and ease of adapting multi-channel to
single-channel data), an axis is added to the output from MainParameters
if only a single channel is processed.  This means that all data will
always have shape (n_channels, n_events) in the DataHandler, even if
n_channels=1 (in which case, this axis was dropped previously).

This matches the way that DataHandler.calc_mp functioned.