bug: Return RMS instead of RMS when model > 0
When model = 0, FitBaseline uses np.std to calculate the RMS of the
baseline.  When model > 0, the square root of the residuals is returned
(as `np.sqrt(err)`, where `err` is residuals returned from
`np.linalg.lstsq`).  This change divides the residuals by the number of
samples, converting the residual sum of squares to the root-mean-square.

As a sanity check, this matches the return value of `np.std` for flat
baselines.