bug: Allow int for `where` argument, and check it is less than 1
The `where` argument to FitBaseline assumes it can only be a float, but
an edge case is when `1` is passed (i.e. as an integer meaning to use
the full trace, instead of `1.0`).  This change allows `where` to be an
integer, but also checks that `where <= 1` (i.e. 1 is the only valid
integer).

Also updates the docstring, since it used to say `where` could be an
integer, but the type annotation indicates only a float.