feat: Add designated TP stream channels
This change allows one to specify one (or more) designated .csmpl files
which have been used to record raw TP pulses (i.e. directly from the
pulser), and to trigger on this stream(s) to find the TP timestamps.
The TPA is determined from a list passed as a second keyword argument.