fix(ecal): Implement bisection for inverting transfer function.
For poorly behaved transfer functions (in particular when extrapolating to higher TPEs) the secant method root search could find undesirable solutions. Now, the TPA (as well as the corresponding values of the transfer function at those points) serve as bin edges into which the values of interest are binned. If successful, the right bin edges are used as the starting interval for a bisection search (guaranteed to find root). Only if no such interval is found (i.e. requested value outside of TPE domain), the secant method is used.