bug: Force model=0 in MainParameters for FQL correction
The onset is incorrectly calculated in edge cases of sharply decaying
baselines with the default arguments to `FitBaseline` used in
`MainParameters`, specifically with `model=1`.  In these cases, the
correction would be imposed well ahead of the actual pulse onset.  This
is fixed with `model=0` (which was the default for `CalcMP`).